pyproject.toml
[tool.poetry]
name = "exercise-library"
version = "0.1.0"
description = ""
authors = ["alstacon <alstacon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
crashtest = "0.4.1"
cryptography = "42.0.5"
defusedxml = "0.7.1"
distlib = "0.3.8"
django-environ = "0.11.2"
django-filter = "24.2"
django-templated-mail = "1.1.1"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
djoser = "2.2.2"
dulwich = "0.21.7"
factory-boy = "3.3.0"
faker = "24.4.0"
fastjsonschema = "2.19.1"
filelock = "3.13.3"
idna = "3.6"
inflection = "0.5.1"
iniconfig = "2.0.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
more-itertools = "10.2.0"
msgpack = "1.0.8"
oauthlib = "3.2.2"
packaging = "24.0"
pexpect = "4.9.0"
pkginfo = "1.10.0"
platformdirs = "4.2.0"
pluggy = "1.4.0"
poetry = "1.8.2"
poetry-core = "1.9.0"
poetry-plugin-export = "1.7.1"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pycparser = "2.22"
pyjwt = "2.8.0"
pyproject-hooks = "1.0.0"
pytest = "8.1.1"
pytest-factoryboy = "2.7.0"
python-dateutil = "2.9.0.post0"
python3-openid = "3.2.0"
rapidfuzz = "3.7.0"
requests = "2.31.0"
requests-oauthlib = "2.0.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
six = "1.16.0"
social-auth-app-django = "5.4.0"
social-auth-core = "4.5.3"
sqlparse = "0.4.4"
tomlkit = "0.12.4"
trove-classifiers = "2024.3.25"
typing-extensions = "4.10.0"
urllib3 = "2.2.1"
virtualenv = "20.25.1"
xattr = "1.1.0"
django = "^5.0.4"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
coverage = "^7.4.4"

[tool.pytest.ini_options]
markers = [
    "django_db: mark a test to use db.",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"